Deep Research Agent/requirements.txt
agno==1.2.6
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
azure-ai-inference==1.0.0b9
azure-core==1.32.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==7.5.5
exa-py==1.10.0
exceptiongroup==1.2.2
executing==2.2.0
feedparser==6.0.11
filelock==3.18.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.35.0
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpickle==4.0.5
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.1
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.3.0
mypy-extensions==1.0.0
narwhals==1.33.0
nest-asyncio==1.6.0
networkx==3.4.2
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==1.26.4
openai==1.70.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
pluggy==1.5.0
primp==0.14.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
pydeck==0.9.1
Pygments==2.19.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyvis==0.3.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
streamlit==1.44.0
streamlit-flow==0.1.0
streamlit-flow-component==1.6.1
tenacity==9.0.0
tldextract==5.1.3
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
wikipedia==1.4.0
yarl==1.18.3
zstandard==0.23.0
